Improved the dependency of usage between headers

+++ 04_ASM2IR/src/CPU.cpp
@@ -9,7 +9,7 @@ void CPU::dumpCPU() const {
   std::cout << "pc: " << std::hex << pc << '\n';
   std::cout << "next_pc: " << next_pc << '\n';
 
-  for (uint32_t reg = 0; reg < reg_size; ++reg)
+  for (uint64_t reg = 0; reg < reg_size; ++reg)
     std::cout << "x" << reg << ": " << reg_file[reg] << '\n';
 
   std::cout << '\n';
